:sparkles: Implement new preview
