+++ app/src/components/Common/TextInput/TextInput.module.scss
@@ -13,7 +13,7 @@
   outline: none;
   border: none;
   border-bottom: solid 1px;
-  border-color: grey;
+  border-color: var(--disabled);
   background-color: var(--backgroundColor);
   display: block;
   padding: 10px 10px 0 10px;
