--- app/src/components/PreviewEvent/PreviewEvent.module.scss
@@ -1,26 +1,0 @@
-@import 'src/style/constants';
-
-.previewContainer {
-  width: 60%;
-}
-
-.dataEntry {
-  display: flex;
-  justify-content: space-between;
-  margin: 20px 0;
-  white-space: pre-wrap;
-  @include regularText;
-}
-
-.header {
-  @include headerText;
-}
-
-.infoText {
-  text-align: right;
-  width: 65%;
-}
-
-.labelText {
-  width: 35%;
-}
