+++ app/src/components/ViewEvent/ViewEventContainer.module.scss
@@ -1,117 +1,11 @@
 @import 'src/style/constants';
 
-.container {
-  width: auto;
-  display: grid;
-  grid-template-columns: repeat(3, auto);
-  gap: 10px;
-  @media #{$desktop} {
-    max-width: 90%;
-  }
-}
-
-.header {
-  @include headerText;
-  grid-column: 1 / 4;
-  grid-row: 1;
-  margin: 25px 0 10px 0;
-  @media #{$desktop} {
-    grid-column: 1 / 3;
-    margin: 25px 0;
-  }
-}
-
-.buttonContainer {
-  grid-column: 1;
-  grid-row: 2;
-  margin-bottom: 10px;
-  @media #{$desktop} {
-    grid-row: 1;
-    grid-column: 3;
-    margin: auto;
-    margin-right: 0;
-  }
-}
-
 .text {
   @include regularText;
   line-height: 1.2;
   margin: 0;
 }
 
-.timeContainer {
-  @extend .text;
-  grid-row: 3;
-  grid-column: 1 / 4;
-  @media #{$desktop} {
-    grid-column: 1 / 2;
-    grid-row: 2;
-  }
-}
-
-.participantsContainer {
-  @extend .text;
-  grid-row: 4;
-  grid-column: 1 / 3;
-  margin-bottom: 10px;
-  @media #{$desktop} {
-    grid-column: 2;
-    grid-row: 2;
-  }
-}
-
-.locationContainer {
-  @extend .text;
-  grid-row: 4;
-  grid-column: 2 / 4;
-  @media #{$desktop} {
-    grid-row: 3;
-  }
-}
-
-.infoHeader {
-  @extend .text;
-  font-weight: bold;
-  margin: 0;
-  margin-bottom: 10px;
-}
-
-.organizerContainer {
-  @extend .text;
-  grid-row: 5;
-  grid-column: 1 / 4;
-  @media #{$desktop} {
-    grid-row: 3;
-    grid-column: 1 / 2;
-  }
-}
-
-.emailLink {
-  @extend .text;
-  display: block;
-  margin-top: 5px;
-}
-
-.descriptionContainer {
-  grid-row: 6;
-  grid-column: 1 / 4;
-}
-
-.textBlock {
-  @extend .text;
-  margin: 20px 0px;
-  white-space: pre-wrap;
-}
-
 .subHeader {
   @include subHeaderText;
-}
-
-.dateText {
-  margin: 0;
-  margin-bottom: 10px;
-  min-width: 100px;
-  @media #{$desktop} {
-    min-width: 320px;
-  }
 }
